+++ src/main/kotlin/study/Introduce.kt
@@ -1,0 +1,6 @@
+package study
+
+fun introduce(block: PersonBuilder.() -> Unit): Person {
+    // also, apply, run, let, with
+    return PersonBuilder().apply(block).build()
+}
